Move load ideas to pages
