+++ lib/modules/apostrophe-pages/index.js
@@ -5,6 +5,7 @@ const moment          = require('moment'); // returns the new locale, in this ca
 const url             = require('url');
 const internalApiUrl  = process.env.INTERNAL_API_URL;
 const cache           = require('../../../services/cache').cache;
+const loadIdeas       = require('./lib/load-ideas');
 
 module.exports = {
     types: [
@@ -59,5 +60,18 @@ module.exports = {
 
         return superPageBeforeSend(req, callback);
       };
+
+      /**
+       * Load in the ideas from the api and set to the req object
+       */
+      self.apos.app.use((req, res, next) => {
+        req.data.csrf = self.apos.settings.getOption(req, 'csrf');
+        req.data.apiUrl = self.apos.settings.getOption(req, 'apiUrl');
+        req.data.appUrl = self.apos.settings.getOption(req, 'appUrl');
+        req.data.googleMapsApiKey =  self.apos.settings.getOption(req, 'googleMapsApiKey');
+        
+        loadIdeas(req, res, next)
+      });
+
     }
 };
